\name{barnlarv-package}
\alias{barnlarv-package}
\alias{barnlarv}
\docType{package}
\title{Comparative outline analysis of barnacle nauplii}
\description{
  This is a companion package to the manuscript:

  Wong, J. Y., Chan, K. Y. K., Chan, B. K. K. (in press). Phylogenetic,
  ecological and biomechanical constraints on larval form: A comparative
  morphological analysis of barnacle nauplii. \emph{PLoS ONE}.

  to reproduce the results. Please refer to readme at Github repository for
  explanation on how to use the package and updates.
  }
