\name{outline}
\docType{data}
\alias{outline}
\alias{stage2outline}
\alias{ontogeny_outline}
\title{Outlines of barnacle naupliar cephalic shield}
\description{
  There are two datasets of cephalic shield outlines:
  \describe{
    \item{\code{stage2outline}}{outlines from 102 species of
      stage II barnacle, gathered from publised description and lab collection.}
    \item{\code{ontogeny_outline}}{outlines from 12 species of barnacle
    nauplii across developmental stages.}
  }
}
\usage{
  data(stage2outline)
  data(ontogeny_outline)}
\format{
  \describe{
    \item{\code{stage2outline}}{a list with 102 matrices containing the xy
      coordinates of outlines}
    \item{\code{ontogeny_outline}}{a list with 67 matrices containing the xy
      coordinates of outlines}
  }
}
\seealso{\code{\link{semilandmark}}}
\keyword{datasets}
